zfs: vdev: 'fix' to vdev children recognition
This patch is a temporary fix to spa load procedure when recognizing
vdev children. I strongly think that it's due to our own versions of
the functions spa_generate_rootconf and spa_import_rootpool which fail
to get proper config data from children vdevs.

The root of the problem could also be that persistent spa config support
was disabled by the commit d0d5e40c96.
But reading the spa load procedure, it turns out that ZFS would fall
back to another approach where the configuration would be made up
from scratch, possibly more the root of the slowness on ZFS mount.

Our ZFS port is still incomplete, and there is probably a bunch of other
problems that must be eventually fixed, e.g. ZFS replay is currently
disabled, and in event of power failure, data will potentially be lost.

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>